requirements.txt
pymongo==4.3.3
python-dotenv==1.0.0
fastapi==0.103.1
bcrypt==4.0.1
